+++ src/testmod/java/dev/onyxstudios/componenttest/TestComponents.java
@@ -132,8 +132,9 @@ public final class TestComponents implements
 
     @Override
     public void registerItemComponentFactories(ItemComponentFactoryRegistry registry) {
+        // this first line adds data to every stack, which is an awful idea for several reasons
         registry.registerFor(i -> true, ALT_VITA, (stack) -> new BaseVita(stack.getItem() == Items.DIAMOND_CHESTPLATE ? 3 : stack.getCount()));
-        registry.registerFor(CardinalComponentsTest.VITA_STICK_ID, VITA, stack -> new BaseVita());
+        registry.registerFor(CardinalComponentsTest.VITALITY_STICK, VITA, stack -> new BaseVita());
     }
 
     @Override